Making some changes to use for icehouse version of heat.
Change the floating network parameter to accept a floating network
UUID instead of name of the network. This uuid is then used
wherever the previoud name of the network would be used.

Fixed the generation of the random string. Moved from providing a
list of character classes to providing just one sequence with the
same character class.